(related to bug 2497) Adjusting variable errors and changing the Angle to float.
git-svn-id: file:///tmp/svnroot_only_rec/internal/rec-project/rec/trunk@9456 8c4c7310-a39f-45c4-bd0d-b10b2e5d99a4

+++ experiments/polaroid/src/java/server/pt/utl/ist/elab/driver/polaroid/processors/StampPolaroidProcessor.java
@@ -42,11 +42,11 @@ public class StampPolaroidProcessor extends AbstractStampProcessor {
 				&& splitedStr[0] != null) {
 			try {
 				angulo = Integer.parseInt(splitedStr[0]);
-				final Integer oAngulo = new Integer((angulo * 180) / 1300);
+				final Float oAngulo = new Float((angulo * 180F) / 1300F);
 				command.addCommandData(StampPolaroidProcessor.ANGULO, oAngulo);
 
 				intensidade = Integer.parseInt(splitedStr[1]);
-				final Float oIntensidade = new Float((intensidade * 5) / 4096F);
+				final Float oIntensidade = new Float((intensidade * 5F) / 4096F);
 				command.addCommandData(StampPolaroidProcessor.INTENSIDADE, oIntensidade);
 
 				command.setData(true);
